add: initialize test
