+++ jest.config.js
@@ -1,0 +1,5 @@
+export default {
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    testMatch: ['**/tests/**/*.test.ts'],
+};
