+++ package.json
@@ -4,15 +4,16 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start:dev": "npx kill-port 3000 5000 9000 && npx nodemon",
     "dev": "node build/index.js",
     "build": "rimraf ./build && tsc",
-    "start": "node api/index.ts",
+    "start": "nodemon src/index.ts",
     "knex": "ts-node -r ./node_modules/dotenv/config ./node_modules/knex/bin/cli.js  --knexfile src/config/knexfile.ts",
     "migrate:make": "ts-node -r ./node_modules/dotenv/config ./node_modules/knex/bin/cli.js  --knexfile src/config/knexfile.ts migrate:make",
     "migrate:latest": "ts-node -r ./node_modules/dotenv/config ./node_modules/knex/bin/cli.js  --knexfile src/config/knexfile.ts migrate:latest",
-    "migrate:rollback": "ts-node -r ./node_modules/dotenv/config ./node_modules/knex/bin/cli.js  --knexfile src/config/knexfile.ts migrate:rollback"
+    "migrate:rollback": "ts-node -r ./node_modules/dotenv/config ./node_modules/knex/bin/cli.js  --knexfile src/config/knexfile.ts migrate:rollback",
+    "test": "cross-env NODE_ENV=test jest --detectOpenHandles",
+    "migrate:test": "cross-env NODE_ENV=test ts-node -r ./node_modules/dotenv/config ./node_modules/knex/bin/cli.js  --knexfile src/config/knexfile.ts migrate:latest"
   },
   "author": "Adira",
   "license": "ISC",
@@ -51,16 +52,22 @@
     "@types/bcrypt": "^5.0.2",
     "@types/body-parser": "^1.19.5",
     "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.14",
     "@types/jsonwebtoken": "^9.0.5",
     "@types/multer": "^1.4.11",
     "@types/node": "^20.10.4",
+    "@types/supertest": "^6.0.3",
     "@types/swagger-ui-dist": "^3.30.4",
     "@types/validator": "^13.11.8",
     "@typescript-eslint/eslint-plugin": "^6.13.2",
     "@typescript-eslint/parser": "^6.13.2",
+    "cross-env": "^7.0.3",
     "eslint": "^8.55.0",
+    "jest": "^29.7.0",
     "nodemon": "^3.0.2",
     "rimraf": "^5.0.5",
+    "supertest": "^7.1.0",
+    "ts-jest": "^29.3.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3"
   }
